Modernize the CMakeLists.txt that use LEDA
The file `${CGAL_USE_FILE}` must be included before
`find_package(LEDA)` can be called. But that would require to verify
`CGAL_FOUND` first...

Actually, now that we require at least CMake 3.1, let's say that CGAL
is `REQUIRED`, but not the components. That is the basic truth, and
that avoids testing for `CGAL_FOUND`.

At the same time, I have modified the `cmake_minimum_required`:
  - placed as the first line,
  - set to `3.1...3.13`:
      https://cmake.org/cmake/help/latest/command/cmake_policy.html#setting-policies-by-cmake-version

I have also removed the comments that says the `CMakeLists.txt` was
generated. That may have been true... but before a lot of modifications.