Fix the default of 3rd template parameter of Convex_hull_traits_3
The documentation of `convex_hull_3` says that, if the kernel is "like
`Epick`", then the default traits class of `convex_hull_3` is
`Convex_hull_traits_3`.

But that is wrong, because `Convex_hull_traits_3` is documented with
only two template parameters, and if the kernel is "like `Epick`", the
third argument is `CGAL::Tag_true`, whereas the default of that
parameter is `CGAL::Tag_false`. If users want to write explicitly the
traits in the call to `convex_hull_3`, they cannot know that there is
a third parameters, and that it must be `Tag_true`.

This commit fixes the default of that third argument.