Strange patch that fixes the remaining bugs
https://github.com/CGAL/cgal/pull/4013#issuecomment-507291311

Because of two constructions in Algebraic_curve_kernel_2 and
Hyperbolic_octagon_translation, the assertion "!blocks.empty()" from
`<CGAL/CORE/MemoryPool.h>` was triggered during the destruction of
thread-local objects.

This strange patch ensures that the order of creation of thread-local
object is right, so the order of destruction is right as well.