Remove that assertion
In a multi-threaded application, the double map can be modified while
the `size()` function is queried. That is a race-condition. We usually
do not care, because the query is here only to display messages, but
we do not want the assertion to trigger because the two sides of the
double map were being modified while another thread calls `is_valid()`.