fix the degenerate self-intersections issue
almost degenerate faces tend to create self-intersections in the output mesh.
With this commit, `fix_degenerate_faces()` now also deals with almost degenerate
faces, namely the faces with their area < 1% * average area of neighbors
(we only compute squared areas though it is not the exact same criterion)

About fix_degenerate_faces().
There also are some configurations that this function is not able to fix :
sometimes, flipping an edge to fix a degenerate face creates a new
degenerate face. And flipping an edge to fix it comes back to the
initial configuration.
After 10 attempts around `v`, we give up and just return true
because the mesh has been modified