Fix compilation errors, with a few compiler
```
include/CGAL/Triangulation_hierarchy_3.h:163:23: error: invalid use of 'this' at top level
     noexcept(noexcept(this->Tr_Base::swap(tr)))
                       ^~~~
```

That is actually the subject of a C++ Defect:
http://www.open-std.org/jtc1/sc22/wg21/docs/cwg_defects.html#1207

Anyway, this was used only for the `noexcept` specification of `swap`
functions, and I no longer use `swap` for the move semantic. I can
remove those noexcept` specifications.