shape detection: Fix plane normals being arbitrary in region growing.
In #3864 region growing was introduced, but the
`linear_least_squares_fitting_3()` used in `update()` cannot
estimate a correct face normal direction from just unoriented vertex points
(the plane normal's orientation signum may be flipped).

This commit flips the estimated plane normal to face into the
same direction as the majority of the faces it is estimated from.

With the estimated planes now being "the right way around",
this commit subsequently also fixes the incorrect use of `abs()`
in `is_part_of_region()` when comparing a face's normal to the
estimated plane normal (direction matters here).

This fixes parallelly-opposing planar regions (means: with opposite normals),
connected with some triangles in between, like this:

      | ->      <- |
      | ->      <- |      Legend:
      | ->      <- |        face:   |
      | ->      <- |        normal: ->
      | ->      <- |
       \__________/

being detected as 1 region.

This should have been impossible according to the docs
("... and if the angle between its normal and the plane's normal is within
the `angle_threshold`"), given the documented requirement
("\pre `angle_threshold >= 0 && angle_threshold <= 90`").
Now it is impossible.