Restore old static filtering mechanisms:
One cannot use static_filters because it extracts a lot of FT
in the call to fit_in_double(), but for FRK, a Point_3 is
a pair<AK::Point_3, EK::Point_3> and so calling p.x() and such
functions is constructing a lot of FT, which is expensive.

Thus, FRK fully relies on EPICK for predicates, even
if sometimes (e.g. in DT3), it means just being a wrapper
around EPICK.