Improving Interface
The "Draw" and "Insert" button sets ("Draw" and "insert" columns) serve the same
purpose. Instead, a single row is added that also indicates which color bucket is currently used for input. It should also be used for loading data from an external file in the future.

The hiding button set (currently, most left) is entitled "Visible" and shift to the rightmost position now it seems to appear far more intuitive than before.

Clearing out a color bucket should is made easier by introducing a new column "Clean" and a function "Clean" which keeps it within the “draw” column makes it more intuitive.