Fix a bug in <CGAL/Constrained_triangulation_2.h>
The function parameter was locally shadowed by a local `int i`
variable!

Then the edge `(f, i)` in the call `remove_constrained_edge(f, i)` was
the wrong one!

The bug was introduced by ecfd82e287378977a325ae34e21087a2075a05cd,
ten years ago.

The bug was only triggered on degenerated cases, when two constraints
intersect, but the code failed to compute the intersection. Then, if
the intersection vertex `vi` was either `vaa` or `vbb` (and not `vcc`
or `vdd`), then the edge `(f, i)` was the wrong one.