Double defined ID seg_coord_global
It is better to define the anchor in the detailed section of a method as when it is in the brief description it will be shown in the overview of the functions (top of the page) and in the detailed description.
This is due to the `REPEAT_BRIE=YES` setting.

See pages Barycentric_coordinates_2/classCGAL_1_1Barycentric__coordinates_1_1Segment__coordinates__2.html and Barycentric_coordinates_2/index.html