Remove (broken) HDT2::insert(p, lt, li) from documentation
The 'lt' is documented to be HDT2::Locate_type, which is wrong, the function's
signature uses HDT2::DT2::Locate_type, and that's for a precise reason:
HDT2::locate() returns location based on hyperbolic simplicies, and not
the underlying DT2 triangulation. So, one might retrieve "OUTSIDE_CONVEX_HULL"
if a query is on a non-hyperbolic Delaunay face, but a finite underlying DT2
face. In that configuration, it'd break the triangulation to do an insertion
with "OUTSIDE_CONVEX_HULL".

The documentation could be fixed, but it is too dangerous so it's best
to just remove it. (Maybe it should also be simply removed
from the triangulation class.)