Polynomial: background of formula in advanced section
In the file `Polynomial/group__PkgPolynomialFunctions.html#ga0148f1612807b08b433f96c3ce07f2be` we see for the function `number_of_real_roots` in the advanced section that all formulas except 1 are nicely shown as (MathJax) formulas and that one is an image of which the background color is incorrect.
Replacing the image with a proper formula makes all consistent and proper.