Possibility to disable special CGAL documentation setting
In the past it was possible by means of the option `-DCGAL_EXTRACT_ALL_NO_DETAILED_IF_EMPTY=OFF` with CMake to disable the specific CGAL setting though currently with the 1.8.13 version this didn't work (anymore).
Also the setting has been renamed for the 1.9.6 version (and newer) to `NO_ADDITIONAL_DETAILS` this has now been made more flexible as well.