Add missing header
CGAL-5.5.x latest test results are completely broken.

https://cgal.geometryfactory.com/CGAL/testsuite/results-5.5.4-I-166.shtml

That is not due to any code change (there was none since the previous testsuite for 5.5.x). The error are due to Boost 1.84. Probably something change in the includes of Boost. The error is about `boost::prior`.  The header `<boost/next_prior.hpp>` must be included.