Use the real circumradius value to sort the facets
Meaning, use the value that we compare against alpha, and not
simply the radius of the smallest circumscribing ball.

This strongly changes the order of the queue and thus thus results
are very different, but still the same (same guarantees, same
element quality, only a little bit more elements, etc.)

Also a massive, ~35% speed-up, that needs to be investigated.