Use a single Lazy_construction class
Within the operator(), we do mostly what was done in separate functors
before, but can now be done in the same function because even
though the return type (deduced by decltype(auto)) is not the same
for each 'if' block, the C++17 constexpr allows different return
types in the different blocks.

Not only we factorize code, but we do not have to use result_types
anymore, and we can get the true result_type thanks to decltype()
on the actual functor + parameters call.