Fix missing some pierce events
Some events like a surface event can increase the combinatorial distance
between a vertex and a facet such that a pierce event becomes
a detectable event, whereas before the has_edge_to_facet filter
would ignore it. The vertex can be unmodified and so it is not part
of the "local" vertices if we do local queue updates.

As to fix this: always allow such events, but delay the the combinatorial
filter at pop time, when we know the event is relevant.

It means we compute 4-plane intersections slightly more often,
but the alternatives are:
- missing pierce events
- needing a lot more "local" vertices when updating the queue locally.

So for now, let's compute a little more until it becomes an issue.