Color overnight documentation build report fields
Currently a line is colored in the overnight documentation build report based on whether or not one of the builds  for a package has a warning or error.
With this proposed patch the respective fields are colored (and the package name), so when for a package one of the builds fails that specific field is colored (as well as the package name).